sabrina-2/pyproject.toml
[project]
name = "sabrina"
version = "0.0.1"
description = "Sabrina AI - personal daily-driver assistant"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "ollama>=0.3",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "rich>=13.7",
    "typer>=0.12",
    "anyio>=4.3",
    # Speaker component
    "sounddevice>=0.4.7",
    "numpy>=1.26",
    "httpx>=0.27",
    # Windows-only: direct SAPI access for the zero-setup TTS fallback.
    "pywin32>=306; sys_platform == 'win32'",
    # Listener component (ASR)
    "faster-whisper>=1.0",
    "soundfile>=0.12",
    # Voice-loop component (push-to-talk hotkey)
    "pynput>=1.7",
    # Settings GUI
    "customtkinter>=5.2",
    "tomlkit>=0.13",
    # Vision component: screen capture + image resize/encode
    "mss>=9.0",
    "pillow>=10.0",
    # Semantic memory: sentence embeddings + sqlite-vec index.
    # sentence-transformers pulls torch (~700MB) transitively. A leaner
    # onnxruntime+MiniLM replacement is on the research list.
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1.6",
]

[project.scripts]
sabrina = "sabrina.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sabrina"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
